* 代码优化

+++ langx-java/src/main/java/com/jn/langx/security/crypto/digest/internal/impl/_KeccakDigest.java
@@ -227,7 +227,7 @@ public class _KeccakDigest implements Digest {
                 KeccakExtract();
                 bitsInQueue = rate;
             }
-            int partialBlock = (int) Math.min((long) bitsInQueue, outputLength - i);
+            int partialBlock = (int) Math.min(bitsInQueue, outputLength - i);
             System.arraycopy(dataQueue, (rate - bitsInQueue) / 8, output, offset + (int) (i / 8), partialBlock / 8);
             bitsInQueue -= partialBlock;
             i += partialBlock;
