+++ langx-java/src/main/java/com/jn/langx/util/hash/streaming/murmur/Murmur3_128Hasher.java
@@ -82,7 +82,7 @@ public class Murmur3_128Hasher extends AdvancedStreamingHasher {
             case 10:
                 k2 ^= (long) toInt(bb.get(9)) << 8; // fall through
             case 9:
-                k2 ^= (long) toInt(bb.get(8)); // fall through
+                k2 ^= toInt(bb.get(8)); // fall through
             case 8:
                 k1 ^= bb.getLong();
                 break;
@@ -99,7 +99,7 @@ public class Murmur3_128Hasher extends AdvancedStreamingHasher {
             case 2:
                 k1 ^= (long) toInt(bb.get(1)) << 8; // fall through
             case 1:
-                k1 ^= (long) toInt(bb.get(0));
+                k1 ^= toInt(bb.get(0));
                 break;
             default:
                 throw new AssertionError("Should never get here.");
